Human_Staging_Portal/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic>=2.6.0
pyyaml==6.0.1
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0
jinja2>=3.1.0
aiofiles>=23.0.0

# Future AI/Research capabilities  
openai>=1.12.0
tavily-python>=0.3.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
requests>=2.31.0 